refactor.canopen: update object dictionary and board configuration
- Move can_node_id to axis configuration
- Add magic number and hardware version to board configuration
- Update PDO mapping for new configuration
- Adjust object dictionary entries for changed parameter types

+++ Inc/axis.hpp
@@ -110,7 +110,7 @@ public:
         LockinConfig_t calibration_lockin = default_calibration();
         LockinConfig_t sensorless_ramp = default_sensorless();
         LockinConfig_t general_lockin;
-        uint32_t can_node_id = 1; // Both axes will have the same id to start
+        uint16_t can_node_id = 1; // Both axes will have the same id to start
         bool can_node_id_extended = false;
         uint32_t can_heartbeat_rate_ms = 100;
         uint32_t offset;
